requirements.txt
numpy
torch
transformers
sympy
matplotlib
scikit-learn
pandas
PyYAML
jupyter